refactor: move Sort getters to Doc
